Remove noise and rename card to index
