--- packages/host/app/controllers/index-card-error.ts
@@ -1,1 +1,0 @@
-export { default } from './card-error';
